LayoutHelper to aid in loading representations' layouts
(mostly for handling prefixes automatically)
I was not sure where to put this, it's location is not firmly set, first I
thought it has to go into RepresentationBase or some other non-templated base
class of it. However not all representations have to load layouts per-se, so
I kept it separate for now.